Added Super Admin functionalities

+++ src/profile/profile.services.ts
@@ -24,13 +24,23 @@ export class ProfileService {
     if(!profile){
       throw new NotFoundException(`User with ${id} not found`)
     }
+   
     return classToPlain(profile);
   }
   async remove(id:any):Promise<{message:string}>{
-    const profile = await this.profileRepository.delete(id)
-    if(profile.affected===0){
+    const profile = await this.profileRepository.findOne({where:{id}})
+
+    if(!profile){
+      throw new NotFoundException(`User with ${id} not found`)
+    }
+    if(profile.role === 'super admin'|| profile.role === 'admin'){
+      throw new NotFoundException(`You cannot delete this profile`)
+    }
+    const profiles = await this.profileRepository.delete(id)
+    if(profiles.affected===0){
       throw new NotFoundException(`User with ID ${id} not found`)
     }
+  
     return {message:`Deleted succesfully`}
   }
